Use double instead of float

+++ packages/react-native-reanimated/Common/cpp/reanimated/Tools/PlatformDepMethodsHolder.h
@@ -14,7 +14,7 @@ using namespace react;
 namespace reanimated {
 
 using SynchronouslyUpdateUIPropsFunction =
-    std::function<void(const std::vector<int> &, const std::vector<float> &)>;
+    std::function<void(const std::vector<int> &, const std::vector<double> &)>;
 using UpdatePropsFunction =
     std::function<void(jsi::Runtime &rt, const jsi::Value &operations)>;
 using ObtainPropFunction = std::function<jsi::Value(
